pyproject.toml
[tool.poetry]
name = "self-driving"
version = "0.0.1"
description = ""
authors = ["Carlos Alvarez <candres.alv@gmail.com>","Cristian Garcia <cgarcia.e88@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
typer = "^0.1.0"
dataget = "^0.4.9"
dicto = "^0.2.0"
pyarrow = "^0.16.0"
flask = "^1.1.1"
pillow = "^7.0.0"
eventlet = "^0.25.1"
plotly = "^4.5.4"
imbalanced-learn = "^0.6.2"
jupyter = "^1.0.0"
matplotlib = "^3.2.1"
opencv-python = "^4.2.0"
# tensorflow-addons = "^0.10.0"
python-socketio = "^4.6.0"
# tensorflow = "^2.2.0"
debugpy = "^1.0.0-beta.9"
tensorflow_hub = "^0.8.0"
tensorflow_probability = "0.10.1"

[tool.poetry.dev-dependencies]
black = {version = "^19.10b0", allow-prereleases = true}
pylint = "^2.4.4"
flake8 = "^3.8.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
